Move a conversion from size_t to int into one single place within BaseDecoderTest.cpp

+++ test/api/BaseDecoderTest.cpp
@@ -70,13 +70,13 @@ void BaseDecoderTest::TearDown() {
 }
 
 
-void BaseDecoderTest::DecodeFrame (const uint8_t* src, int sliceSize, Callback* cbk) {
+void BaseDecoderTest::DecodeFrame (const uint8_t* src, size_t sliceSize, Callback* cbk) {
   uint8_t* data[3];
   SBufferInfo bufInfo;
   memset (data, 0, sizeof (data));
   memset (&bufInfo, 0, sizeof (SBufferInfo));
 
-  DECODING_STATE rv = decoder_->DecodeFrame2 (src, sliceSize, data, &bufInfo);
+  DECODING_STATE rv = decoder_->DecodeFrame2 (src, (int) sliceSize, data, &bufInfo);
   ASSERT_TRUE (rv == dsErrorFree);
 
   if (bufInfo.iBufferStatus == 1 && cbk != NULL) {
